OG Art, Fire Points, New Waves, Wave order
Created orginal art in Asprite 3 viarents saving some in the assets while two in the main the red and yellow. I tweaked the fire points so they were on the correct updated sprites gun and added new code so it creates two bullets. I modifyied the waves to be better, and changed the order. I still need to tweak more of it.